Added theme compare before copying assets
Added a check to compare the current theme's absolute path with the default theme's absolute path.  If they are the same path then only copy the default theme's files to the output folder and skip the unnecessary copy of the current theme.  This resolved an issue with antivirus software occasionally locking the copied files after the first copy and causing the second unnecessary copy to fail.