Change the Markdown parser
Potential security issues have been reported to the marked, but there
is no plan to still be fixed. Because YUIDoc is a one of the development
tool, I have thought unlikely to be affected by the problem. However,
marked is no longer actively maintained, and I'd like to choice a parser
that are more maintenance.

Since YUIDoc only have utilized simply marked as a simple Markdown parser,
change can often be reduced.