add null check
The reason for adding this null check is that sometimes, for eg. in case of building yuidoc for https://github.com/photonstorm/phaser, when you try to toggle from 'classes' panel to 'modules' panel, the following error is being thrown -:

```Uncaught TypeError: Cannot read property 'toLowerCase' of null```.

For catching this error, I think adding a null check would be a good solution. Although I think, there might be some other reason which is leading to the error in the above case. Please check this. For your reference, version of my yuidoc is 0.9.0 .